Move directions into specific submodule
This refactors the maps module to extract out directions.
