+++ gmaps/__init__.py
@@ -3,6 +3,7 @@ from ._version import version_info, __version__  # noqa
 from .maps import *  # noqa
 from .marker import *  # noqa
 from .heatmap import *  # noqa
+from .directions import *  # noqa
 
 
 def _jupyter_nbextension_paths():
